# frontend/src/types/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "人生草稿本 API",
    "description": "Life Draft Book — Interactive narrative game API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/presets": {
      "post": {
        "tags": [
          "Presets"
        ],
        "summary": "Create Preset",
        "description": "Save a character preset.",
        "operationId": "create_preset_api_presets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Presets"
        ],
        "summary": "List Presets",
        "description": "List character presets for the current user.",
        "operationId": "list_presets_api_presets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PresetInfo"
                  },
                  "title": "Response List Presets Api Presets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/presets/{preset_id}": {
      "get": {
        "tags": [
          "Presets"
        ],
        "summary": "Get Preset",
        "description": "Get a single character preset.",
        "operationId": "get_preset_api_presets__preset_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Presets"
        ],
        "summary": "Delete Preset",
        "description": "Delete a character preset.",
        "operationId": "delete_preset_api_presets__preset_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePresetRequest": {
        "properties": {
          "preset_name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Preset Name"
          },
          "player_name": {
            "type": "string",
            "maxLength": 50,
            "title": "Player Name"
          },
          "life_vision": {
            "type": "string",
            "maxLength": 500,
            "title": "Life Vision",
            "default": ""
          },
          "character_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Character Settings",
            "x-maxBytes": 262144
          }
        },
        "type": "object",
        "required": [
          "preset_name",
          "player_name"
        ],
        "title": "CreatePresetRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "PresetInfo": {
        "properties": {
          "preset_id": {
            "type": "integer",
            "title": "Preset Id"
          },
          "preset_name": {
            "type": "string",
            "title": "Preset Name"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "life_vision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Life Vision"
          },
          "character_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Character Settings"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "preset_id",
          "preset_name",
          "player_name"
        ],
        "title": "PresetInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}